Add master mode.
This commit allows us to launch pg_keeper even on master server as
master mode. When pg_keeper is running as master mode, it pool to
standby server which is connected as synchronous standby server in
order to that master server prevent to wait for unconnected synchronous
standby server commit forever.

Regarding this change, source code is invided into two files.