Support multiple stanby node mode on standby side.
We introduce new functions such as indirect_polling(), indirect_kill()
which are used for idirectly polling and updating standby's local cache.

In most case, we don't want to promote any standby server as long as
the master server seems to be live. If the standby server checks connection
to master only from itself, it would not be work fine in case of network failuer
between master server and that standby server. So we introduce indirectly polling
mechanism which asks other standby server to poll the master server. If one
standby can not connect to master but other standby says we can connect to master
server, it's OK. On the other hand if the all stnadbys say that we could
not connect to the master server, it's time to promote.

The master server can update its local cache information when adding new
node, but standby server could not know that management table is updated.
So we introduce indirectly sending signal function which sends SIGUSR1 signal
to pg_keeper process via SQL. Using this function, we can probagate the signal
to all other standbys.

Also we change the process title so that it has the number of standbys on its
local cache.

Masahiko Sawada